% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maspgDist.R
\name{maspgDist}
\alias{maspgDist}
\title{MASPG metric function}
\usage{
maspgDist(
  tree.a,
  tree.b,
  coord_name = "Location",
  discretization = FALSE,
  geography = TRUE
)
}
\arguments{
\item{tree.a}{an object of the class \code{treedata}}

\item{tree.b}{an object of the class \code{treedata} (with the same tip labels as tree.a).}

\item{coord_name}{(optional) a string specifying the attribute name used to indicate the geographic trait variable in the \code{treedata} object.}

\item{discretization}{(optional) I_MASPG is a discrete measure which compares if locations are the same or different, and which does not use geographic distances. To use continuous distances a discretization must be specified. This is a numeric measure of distance under which two points are 'the same'. If discretization is used, 2-dimensional co-ordinates are assumed, with names coord_name1 and coord_name2.}

\item{geography}{(optional) If false, then the Maximum Agreement Sub-Tree (MAST) metric is computed instead of the MASPG metric}
}
\value{
The distance between the two trees according to the metric
}
\description{
Comparison of two phylogeographic trees using the Maximum Agreement Sub-Phylogeography (MASPG) incompatibility metric.
}
\author{
Benjamin John Singer \email{benjamin.singer@bnc.ox.ac.uk}

Antonello Di Nardo \email{antonello.dinardo@pirbright.ac.uk}

Luca Ferretti \email{luca.ferretti@gmail.com}
}
