% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrcaDist.R
\name{mrcaVec}
\alias{mrcaVec}
\title{MRCA vector function}
\usage{
mrcaVec(tree, coord_name = "Location", continuous = FALSE, mrca_matrix = NULL)
}
\arguments{
\item{tree}{a tree object of the class \code{treedata}}

\item{coord_name}{(optional) a string specifying the attribute name used to indicate the geographic trait variable in the \code{treedata} object.}

\item{continuous}{(optional) logical specifying whether to expect discrete location trait (FALSE) or 2-dimensional geographic coordinates (TRUE).}

\item{mrca_matrix}{(optional) the matrix of most recent common ancestors of the tips of \code{tree} as generated by \code{linearMrca(tree@phylo)} function of the \code{treespace} package}
}
\description{
Function to find the location of the most recent common ancestor (MRCA) of each pair of tips in a tree.
}
\author{
Benjamin John Singer \email{benjamin.singer@bnc.ox.ac.uk}

Antonello Di Nardo \email{antonello.dinardo@pirbright.ac.uk}

Luca Ferretti \email{luca.ferretti@gmail.com}
}
