% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maspgDist.R
\name{multiMaspgDist}
\alias{multiMaspgDist}
\title{Metric function for \code{multiPhylo} input}
\usage{
multiMaspgDist(
  trees,
  coord_name = "Location",
  discretization = FALSE,
  geography = TRUE
)
}
\arguments{
\item{trees}{an object of the class \code{multiPhylo} containing the trees to be compared}

\item{coord_name}{(optional) a string specifying the attribute name used to indicate the geographic trait variable in the \code{treedata} object.}

\item{discretization}{(optional) I_MASPG is a discrete measure which compares if locations are the same or different, and which does not use geographic distances. To use continuous distances a discretization must be specified. This is a numeric measure of distance under which two points are 'the same'. If discretization is used, 2-dimensional co-ordinates are assumed, with names coord_name1 and coord_name2.}

\item{geography}{(optional) If false, then the Maximum Agreement Sub-Tree (MAST) metric is computed instead of the MASPG metric.}
}
\value{
The pairwise tree incompatibility matrix
}
\description{
Comparison of a list of phylogeographic trees using the Maximum Agreement Sub-Phylogeography (MASPG) incompatibility metric. Output is given as a pairwise distance matrix.
}
\author{
Michelle Kendall \email{michelle.louise.kendall@gmail.com}

Benjamin John Singer \email{benjamin.singer@bnc.ox.ac.uk}

Antonello Di Nardo \email{antonello.dinardo@pirbright.ac.uk}

Luca Ferretti \email{luca.ferretti@gmail.com}
}
