% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrcaDist.R
\name{multiMrcaDist}
\alias{multiMrcaDist}
\title{Metric function for \code{multiPhylo} input}
\usage{
multiMrcaDist(
  trees,
  coord_name = "Location",
  continuous = FALSE,
  distance.matrix = NULL,
  normalization = TRUE,
  save.memory = FALSE
)
}
\arguments{
\item{trees}{an object of the class \code{multiPhylo} containing the trees to be compared}

\item{coord_name}{(optional) a string specifying the attribute name used to indicate the geographic trait variable in the \code{treedata} object.}

\item{continuous}{(optional) logical specifying whether to expect discrete location trait (FALSE) or 2-dimensional geographic coordinates (TRUE).}

\item{distance.matrix}{(optional) a matrix of distances between locations as stored in the data of \code{tree.a} and \code{tree.b}. If not provided, and continuous = FALSE, then the distance between all locations is set to 1.}

\item{normalization}{(optional) logical specifiying whether distances should be normalised to account for the tree size.}

\item{save.memory}{(optional) logical specifying whether MRCA vectors should be created newly for each comparison.}
}
\value{
The distance between the two trees according to the metric
}
\description{
Comparison of a list of phylogeographic trees using the Pairwise Most-recent Common Ancestor (MRCA) incompatibility metric. Output is given as a pairwise distance matrix.
}
\author{
Michelle Kendall \email{michelle.louise.kendall@gmail.com}

Benjamin John Singer \email{benjamin.singer@bnc.ox.ac.uk}

Antonello Di Nardo \email{antonello.dinardo@pirbright.ac.uk}

Luca Ferretti \email{luca.ferretti@gmail.com}
}
